Rename EventListener to EventSubscriber
Move src/EventListener/ to src/EventSubscriber/ and update the matching
namespaces, the services/event_subscriber.xml config file, and the
top-level services.xml import. Both classes are tagged kernel.event_subscriber,
so the directory name now matches what they actually are.

Also tidy AdminMenuListener with the null-safe operator now that the loops
do nothing else with the child.
